Moving test to nightly, removing ckpt files
