+++ tools/ci_build/build.py
@@ -1280,6 +1280,19 @@ def run_training_python_frontend_tests(cwd):
     run_subprocess([sys.executable, '-m', 'pytest', '-sv', 'orttraining_test_orttrainer_bert_toy_onnx.py'], cwd=cwd)
 
 
+def run_checkpointing_tests(cwd):
+    log.info("Running multi-GPU checkpointing tests.")
+
+    import torch
+    ngpus = torch.cuda.device_count()
+    if ngpus > 1:
+        # generate checkpoint files required in orttraining_test_checkpoint_aggregation.py
+        run_subprocess(['mpirun', '-n', str(ngpus), '-x', 'NCCL_DEBUG=INFO', sys.executable,
+                        'orttrainer_bert_toy_onnx_ckpt_gen.py'], cwd=cwd)
+
+        run_subprocess([sys.executable, '-m', 'pytest', '-sv', 'orttraining_test_checkpoint_aggregation.py'], cwd=cwd)
+
+
 def run_training_python_frontend_e2e_tests(cwd):
     # frontend tests are to be added here:
     log.info("Running python frontend e2e tests.")
@@ -1298,17 +1311,6 @@ def run_training_python_frontend_e2e_tests(cwd):
         run_subprocess([
             'mpirun', '-n', str(ngpus), '-x', 'NCCL_DEBUG=INFO', sys.executable,
             bert_pretrain_script, 'ORTBertPretrainTest.test_pretrain_convergence'], cwd=cwd)
-
-        # TODO: testing in e2e pipeline until multi-node CI is available.
-        log.debug('RUN: mpirun -n {} ''-x' 'NCCL_DEBUG=INFO'' {} {} {}'.format(
-            ngpus, sys.executable, bert_pretrain_script, 'ORTBertPretrainTest.test_pretrain_zero'))
-        run_subprocess([
-            'mpirun', '-n', str(ngpus), '-x', 'NCCL_DEBUG=INFO', sys.executable,
-            bert_pretrain_script, 'ORTBertPretrainTest.test_pretrain_zero'], cwd=cwd)
-
-        # regenerate checkpoint files required in orttraining_test_orttrainer_bert_toy_onnx.py
-        run_subprocess(['mpirun', '-n', str(ngpus), '-x', 'NCCL_DEBUG=INFO', sys.executable,
-                        'orttrainer_bert_toy_onnx_ckpt_gen.py'], cwd=cwd)
 
         log.debug('RUN: mpirun -n {} {} orttraining_run_glue.py'.format(ngpus, sys.executable))
         run_subprocess([
